refactor: move calendar services function in its own file
