+++ src/components/calender/services/get-first-day-of-month.ts
@@ -1,0 +1,3 @@
+export function getFirstDayOfMonth(year: number, month: number): number {
+  return new Date(year, month, 1).getDay()
+}
