+++ src/components/calender/services/get-last-day-of-month.ts
@@ -1,0 +1,3 @@
+export function getLastDayOfMonth(year: number, month: number): number {
+  return new Date(year, month + 1, 0).getDate()
+}
