+++ src/components/calender/services/get-prev-month-and-year.ts
@@ -1,0 +1,5 @@
+export function getPrevMonthAndYear(year: number, month: number): [number, number] {
+  const prevMonth = month - 1
+  const prevYear = prevMonth < 0 ? year - 1 : year
+  return [prevYear, (prevMonth + 12) % 12]
+}
